+++ EzyBook/Application/Info.plist
