--- EzyBook/Core/Data/DTO/Auth/AuthResponseDTO.swift
@@ -1,22 +1,0 @@
-//
-//  AuthResponseDTO.swift
-//  EzyBook
-//
-//  Created by youngkyun park on 5/18/25.
-//
-
-import Foundation
-
-struct AuthResponseDTO: Decodable, EntityConvertible {
-    let accessToken: String
-    let refreshToken: String
-}
-
-extension AuthResponseDTO {
-    func toEntity() -> RefreshEntity {
-        return RefreshEntity(
-            accessToken: self.accessToken,
-            refreshToken: self.refreshToken
-        )
-    }
-}
