+++ EzyBook/Core/Data/Repository/Auth/DefaultAuthNetworkRepository.swift
@@ -9,7 +9,6 @@ import Foundation
 
 /// 실제 네트워크 통신 진행
 final class DefaultAuthNetworkRepository: AuthNetworkRepository {
-
     
     // 통신 및 갱신 Flow 진행
     private let tokenManager: TokenManager
