+++ EzyBook/Data/DTO/Auth/AuthResponseDTO.swift
@@ -1,0 +1,13 @@
+//
+//  AuthResponseDTO.swift
+//  EzyBook
+//
+//  Created by youngkyun park on 5/18/25.
+//
+
+import Foundation
+
+struct AuthResponseDTO: Decodable, EntityConvertible {
+    let accessToken: String
+    let refreshToken: String
+}
