+++ EzyBook/Data/Mapper/Auth/AuthResponseMapper.swift
@@ -1,0 +1,18 @@
+//
+//  AuthResponseMapper.swift
+//  EzyBook
+//
+//  Created by youngkyun park on 5/20/25.
+//
+
+import Foundation
+
+// MARK: AuthResponse Mapper
+extension AuthResponseDTO {
+    func toEntity() -> RefreshEntity {
+        return RefreshEntity(
+            accessToken: self.accessToken,
+            refreshToken: self.refreshToken
+        )
+    }
+}
