+++ EzyBook/Presentation/Authentication/EmailAuth/Models/DisplayError.swift
@@ -1,0 +1,21 @@
+//
+//  DisplayError.swift
+//  EzyBook
+//
+//  Created by youngkyun park on 5/20/25.
+//
+
+import Foundation
+
+enum DisplayError {
+    case error(code: Int, msg: String)
+    
+    var message: (title: String, msg: String) {
+        switch self {
+        case let .error(code, msg):
+            let title = "Error: \(code)"
+            return (title, msg)
+        }
+        
+    }
+}
